pyproject.toml
[tool.poetry]
name = "stealable-landmarks"
version = "0.1.0"
description = ""
authors = ["Samuel Panek <samuelrpanek@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.1"
black = "^21.12b0"
flake8 = "^4.0.1"
docker = "^5.0.3"
djangorestframework-simplejwt = "^5.0.0"
djangorestframework = "^3.13.1"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
